update all bg and fg scripts

+++ k8s2-tags/background.sh
@@ -1,12 +1,9 @@
 #!/bin/bash
-# mkdir k8s-yaml-files
-apt install datamash
-#launch.sh
 touch status.txt
+echo "">/root/status.txt
+if [ ! -f "/root/provisioned" ]; then
+  apt install datamash
+fi
 
-# wall -n "Starting services"
-# kubectl apply -f k8s-yaml-files/lotsofpods.yaml
-# kubectl apply -f k8s-yaml-files/datadog.yaml
-
-# echo "complete">>/root/status.txt
+echo "complete">>/root/status.txt
